chore(notifications): moved notification handlers and event listeners
Moved to a better namespace for code organization
