+++ engine/classes/Elgg/Notifications/Handlers/RemoveAdminUser.php
@@ -1,15 +1,16 @@
 <?php
 
-namespace Elgg\Notifications;
+namespace Elgg\Notifications\Handlers;
 
 use Elgg\Database\QueryBuilder;
+use Elgg\Notifications\NonConfigurableNotificationEventHandler;
 
 /**
  * Notification Event Handler for 'user' 'user' 'remove_admin' action
  *
  * @since 4.0
  */
-class RemoveAdminUserEventHandler extends NonConfigurableNotificationEventHandler {
+class RemoveAdminUser extends NonConfigurableNotificationEventHandler {
 	
 	/**
 	 * Tells if the recipient is the user being changed
