+++ mod/blog/elgg-plugin.php
@@ -168,7 +168,7 @@ return [
 		'object' => [
 			'blog' => [
 				'publish' => PublishBlogEventHandler::class,
-				'mentions' => \Elgg\Notifications\MentionsEventHandler::class,
+				'mentions' => \Elgg\Notifications\Handlers\Mentions::class,
 			],
 		],
 	],
