+++ mod/thewire/elgg-plugin.php
@@ -139,7 +139,7 @@ return [
 		'object' => [
 			'thewire' => [
 				'create' => CreateTheWireEventHandler::class,
-				'mentions' => \Elgg\Notifications\MentionsEventHandler::class,
+				'mentions' => \Elgg\Notifications\Handlers\Mentions::class,
 			],
 		],
 	],
